Migrate `Long` IDs to `Int` in delete and update note routes for consistency with service methods.

+++ src/main/kotlin/io/heapy/kotbusta/ktor/routes/notes/AddOrUpdateNoteRoute.kt
@@ -25,7 +25,7 @@ fun Route.addOrUpdateNoteRoute() {
 
     post("/books/{id}/notes") {
         requireApprovedUser {
-            val bookId = call.requiredParameter<Long>("id")
+            val bookId = call.requiredParameter<Int>("id")
             val request = call.receive<NoteRequest>()
             val note = transactionProvider.transaction(READ_WRITE) {
                 userService.addOrUpdateNote(
